purge the grossness
